+++ src/events/ready.ts
@@ -1,0 +1,12 @@
+import { logger } from '@/core';
+import { Event } from '@/types';
+
+const event: Event = {
+    name: 'ready',
+    once: true,
+    execute(client: any) {
+        logger.info(`Client ready as ${client.user.tag}`);
+    },
+};
+
+export default event;
